+++ bc/cases/tests/test_models.py
@@ -1,0 +1,10 @@
+from unittest import skip
+
+from django.test import TestCase
+
+
+# FIXME this needs to work
+@skip
+class CaseFormPageTest(TestCase):
+    def test_page_loads_when_client_not_configured(self):
+        pass