+++ bc/recruitment/models.py
@@ -5,7 +5,7 @@ from urllib.parse import urlsplit, urlunsplit
 from django.core.exceptions import ValidationError
 from django.db import models
 from django.db.models import Count, F
-from django.db.models.signals import pre_save
+from django.db.models.signals import pre_delete, pre_save
 from django.dispatch import receiver
 from django.http import Http404
 from django.shortcuts import get_object_or_404, render
@@ -187,10 +187,20 @@ class TalentLinkJob(models.Model):
         max_length=255, help_text="Salary group for filtering"
     )
     searchable_location = models.CharField(max_length=255)
+    location_postcode = models.CharField(max_length=8, blank=True)
+    location_lat = models.DecimalField(
+        max_digits=9, decimal_places=6, null=True, blank=True
+    )
+    location_lon = models.DecimalField(
+        max_digits=9, decimal_places=6, null=True, blank=True
+    )
     is_published = models.BooleanField(default=True)
     posting_start_date = models.DateTimeField()
     posting_end_date = models.DateTimeField()
     show_apply_button = models.BooleanField(default=True)
+    attachments = models.ManyToManyField(
+        "documents.CustomDocument", blank=True, related_name="jobs"
+    )
     application_url_query = models.CharField(max_length=255)
 
     def get_categories_list(self):
@@ -218,6 +228,14 @@ class TalentLinkJob(models.Model):
         base_url = self.homepage.url + self.homepage.reverse_subpage("apply")
         scheme, netloc, path, query, fragment = urlsplit(base_url)
         return urlunsplit((scheme, netloc, path, self.application_url_query, fragment))
+
+
+@receiver(pre_delete, sender=TalentLinkJob)
+def callback_talentlinkjob_delete_attachments(sender, instance, *args, **kwargs):
+    # if instance.attachments:
+    for doc in instance.attachments.all():
+        if doc.jobs.all().count() == 1:
+            doc.delete()
 
 
 class RecruitmentHomePage(RoutablePageMixin, BasePage):