+++ CHANGELOG.md
@@ -2,9 +2,20 @@
 
 ## Unreleased
 
+- 114 StreamField Button Link block (BE)
+- 125 Job attachments
 - 138 Additional job sidebar fields (FE)
 - 139 Additional job sidebar fields (BE)
+- 148 Delete outdated jobs when running import command
+
+## 9.0 (2020-03-31)
+
+Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/8.0...9.0
+
+- 129 Jobs postcode search
 - 169 Adjust position of form field help text
+- Forms integration with Aptean Respond API
+- Fix an issue with CI deployments
 
 ## 8.0 (2020-03-30)